+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Development dependencies install via pip: `pip install "merlinwf[dev]"`
+- `merlin status <yaml spec>` that returns queues, number of connected
+  workers and number of unused tasks in each of those queues
 
 ### Fixed
 - `MANIFEST.in` fixes as required by Spack.